chore(arch): Refactor typo in package `quries` into `queries`
