+++ app/src/main/kotlin/com/dot/gallery/feature_node/data/data_source/mediastore/queries/MediaUriFlow.kt
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.dot.gallery.feature_node.data.data_source.mediastore.quries
+package com.dot.gallery.feature_node.data.data_source.mediastore.queries
 
 import android.content.ContentResolver
 import android.content.ContentUris
 import android.database.Cursor
+import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import android.provider.MediaStore
@@ -26,28 +27,30 @@ import com.dot.gallery.core.util.join
 import com.dot.gallery.feature_node.data.data_source.mediastore.MediaQuery
 import com.dot.gallery.feature_node.domain.model.Media
 import com.dot.gallery.feature_node.domain.model.MediaType
+import com.dot.gallery.feature_node.domain.util.isTrashed
 import com.dot.gallery.feature_node.presentation.util.getDate
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+import kotlin.random.Random
 
 /**
- * Media flow
+ * Media uri flow
  *
- * This class is responsible for fetching media from the media store
+ * This class is responsible for fetching media from the media store based on the provided uris
  *
  * @property contentResolver
- * @property buckedId
  * @property mimeType
+ * @property uris
+ * @property reviewMode
  */
-class MediaFlow(
+class MediaUriFlow(
     private val contentResolver: ContentResolver,
-    private val buckedId: Long,
-    private val mimeType: String? = null
+    private val mimeType: String? = null,
+    private val uris: List<Uri>,
+    private val reviewMode: Boolean = false
 ) : QueryFlow<Media.UriMedia>() {
-    init {
-        assert(buckedId != MediaStoreBuckets.MEDIA_STORE_BUCKET_PLACEHOLDER.id) {
-            "MEDIA_STORE_BUCKET_PLACEHOLDER found"
-        }
-    }
+
+    private var buckedId: Long = MediaStoreBuckets.MEDIA_STORE_BUCKET_TIMELINE.id
 
     override fun flowCursor(): Flow<Cursor?> {
         val uri = MediaQuery.MediaStoreFileUri
@@ -57,11 +60,7 @@ class MediaFlow(
                 MediaType.IMAGE -> MediaQuery.Selection.image
                 MediaType.VIDEO -> MediaQuery.Selection.video
             }
-        } ?: when (buckedId) {
-            MediaStoreBuckets.MEDIA_STORE_BUCKET_PHOTOS.id -> MediaQuery.Selection.image
-            MediaStoreBuckets.MEDIA_STORE_BUCKET_VIDEOS.id -> MediaQuery.Selection.video
-            else -> MediaQuery.Selection.imageOrVideo
-        }
+        } ?: MediaQuery.Selection.imageOrVideo
         val albumFilter = when (buckedId) {
             MediaStoreBuckets.MEDIA_STORE_BUCKET_FAVORITES.id -> MediaStore.Files.FileColumns.IS_FAVORITE eq 1
 
@@ -126,46 +125,68 @@ class MediaFlow(
         )
     }
 
-    override fun flowData() = flowCursor().mapEachRow(MediaQuery.MediaProjection) { it, indexCache ->
-        var i = 0
+    override fun flowData() =
+        flowCursor().mapEachRow(MediaQuery.MediaProjection) { it, indexCache ->
+            var i = 0
 
-        val id = it.getLong(indexCache[i++])
-        val path = it.getString(indexCache[i++])
-        val relativePath = it.getString(indexCache[i++])
-        val title = it.getString(indexCache[i++])
-        val albumID = it.getLong(indexCache[i++])
-        val albumLabel = it.tryGetString(indexCache[i++], Build.MODEL)
-        val takenTimestamp = it.tryGetLong(indexCache[i++])
-        val modifiedTimestamp = it.getLong(indexCache[i++])
-        val duration = it.tryGetString(indexCache[i++])
-        val size = it.getLong(indexCache[i++])
-        val mimeType = it.getString(indexCache[i++])
-        val isFavorite = it.getInt(indexCache[i++])
-        val isTrashed = it.getInt(indexCache[i++])
-        val expiryTimestamp = it.tryGetLong(indexCache[i])
-        val contentUri = if (mimeType.contains("image"))
-            MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-        else
-            MediaStore.Video.Media.EXTERNAL_CONTENT_URI
-        val uri = ContentUris.withAppendedId(contentUri, id)
-        val formattedDate = modifiedTimestamp.getDate(Constants.FULL_DATE_FORMAT)
-        Media.UriMedia(
-            id = id,
-            label = title,
-            uri = uri,
-            path = path,
-            relativePath = relativePath,
-            albumID = albumID,
-            albumLabel = albumLabel ?: Build.MODEL,
-            timestamp = modifiedTimestamp,
-            takenTimestamp = takenTimestamp,
-            expiryTimestamp = expiryTimestamp,
-            fullDate = formattedDate,
-            duration = duration,
-            favorite = isFavorite,
-            trashed = isTrashed,
-            size = size,
-            mimeType = mimeType
-        )
-    }
+            val id = it.getLong(indexCache[i++])
+            val path = it.getString(indexCache[i++])
+            val relativePath = it.getString(indexCache[i++])
+            val title = it.getString(indexCache[i++])
+            val albumID = it.getLong(indexCache[i++])
+            val albumLabel = it.tryGetString(indexCache[i++], Build.MODEL)
+            val takenTimestamp = it.tryGetLong(indexCache[i++])
+            val modifiedTimestamp = it.getLong(indexCache[i++])
+            val duration = it.tryGetString(indexCache[i++])
+            val size = it.getLong(indexCache[i++])
+            val mimeType = it.getString(indexCache[i++])
+            val isFavorite = it.getInt(indexCache[i++])
+            val isTrashed = it.getInt(indexCache[i++])
+            val expiryTimestamp = it.tryGetLong(indexCache[i])
+            val contentUri = if (mimeType.contains("image"))
+                MediaStore.Images.Media.EXTERNAL_CONTENT_URI
+            else
+                MediaStore.Video.Media.EXTERNAL_CONTENT_URI
+            val uri = ContentUris.withAppendedId(contentUri, id)
+            val formattedDate = modifiedTimestamp.getDate(Constants.FULL_DATE_FORMAT)
+            Media.UriMedia(
+                id = id,
+                label = title,
+                uri = uri,
+                path = path,
+                relativePath = relativePath,
+                albumID = albumID,
+                albumLabel = albumLabel ?: Build.MODEL,
+                timestamp = modifiedTimestamp,
+                takenTimestamp = takenTimestamp,
+                expiryTimestamp = expiryTimestamp,
+                fullDate = formattedDate,
+                duration = duration,
+                favorite = isFavorite,
+                trashed = isTrashed,
+                size = size,
+                mimeType = mimeType
+            )
+        }.let { flow ->
+            val ids = uris.map {
+                try {
+                    ContentUris.parseId(it)
+                } catch (e: NumberFormatException) {
+                    Random.nextInt(1000000, 2000000)
+                }
+            }
+            if (reviewMode) {
+                flow.map { mediaList ->
+                    mediaList.filter { media ->
+                        ids.contains(media.id) && !media.isTrashed
+                    }
+                }
+            } else {
+                flow.map { mediaList ->
+                    mediaList.filter { media ->
+                        ids.contains(media.id)
+                    }
+                }
+            }
+        }
 }
